requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.4.1; python_version >= '3.6'

certifi==2021.5.30

cffi==1.14.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

cryptography==3.4.7

defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

dj-database-url==0.5.0

django-allauth==0.44.0

django-cors-headers==3.7.0

django-heroku==0.3.1

django-rest-auth[with_social]==0.9.5

django==3.2.5

djangorestframework-simplejwt==4.7.2

djangorestframework==3.12.4

environs==9.3.2

gunicorn==20.1.0

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

marshmallow==3.12.1; python_version >= '3.5'

oauthlib==3.1.1; python_version >= '3.6'

psycopg2-binary==2.9.1

psycopg2==2.9.1

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyjwt[crypto]==2.1.0; python_version >= '3.6'

python-decouple==3.4

python-dotenv==0.18.0

python3-openid==3.2.0

pytz==2021.1

requests-oauthlib==1.3.0

requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.4.1; python_version >= '3.5'

urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

whitenoise==5.2.0
pillow = "*"


